+++ src/main/java/com/phoenix/amazon/AmazonBackend/controllers/impl/CategoryController.java
@@ -1,9 +1,9 @@
 package com.phoenix.amazon.AmazonBackend.controllers.impl;
 
 import com.phoenix.amazon.AmazonBackend.controllers.ICategoryController;
-import com.phoenix.amazon.AmazonBackend.dto.ApiResponse;
-import com.phoenix.amazon.AmazonBackend.dto.CategoryDto;
-import com.phoenix.amazon.AmazonBackend.dto.PageableResponse;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.CategoryDto;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.ApiResponse;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.PageableResponse;
 import com.phoenix.amazon.AmazonBackend.exceptions.BadApiRequestExceptions;
 import com.phoenix.amazon.AmazonBackend.exceptions.CategoryExceptions;
 import com.phoenix.amazon.AmazonBackend.exceptions.CategoryNotFoundExceptions;
