+++ src/main/java/com/phoenix/amazon/AmazonBackend/dto/requestDtos/CategoryDto.java
@@ -1,4 +1,4 @@
-package com.phoenix.amazon.AmazonBackend.dto;
+package com.phoenix.amazon.AmazonBackend.dto.requestDtos;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
