+++ src/main/java/com/phoenix/amazon/AmazonBackend/dto/requestDtos/ProductDto.java
@@ -1,4 +1,4 @@
-package com.phoenix.amazon.AmazonBackend.dto;
+package com.phoenix.amazon.AmazonBackend.dto.requestDtos;
 
 import java.time.LocalDateTime;
