+++ src/main/java/com/phoenix/amazon/AmazonBackend/dto/responseDtos/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.phoenix.amazon.AmazonBackend.dto;
+package com.phoenix.amazon.AmazonBackend.dto.responseDtos;
 
 import org.springframework.http.HttpStatus;
