+++ src/main/java/com/phoenix/amazon/AmazonBackend/dto/responseDtos/ErrorDetails.java
@@ -1,4 +1,4 @@
-package com.phoenix.amazon.AmazonBackend.dto;
+package com.phoenix.amazon.AmazonBackend.dto.responseDtos;
 
 import java.time.LocalTime;
 import java.util.Map;
