+++ src/main/java/com/phoenix/amazon/AmazonBackend/services/IUserService.java
@@ -1,10 +1,10 @@
 package com.phoenix.amazon.AmazonBackend.services;
 
-import com.phoenix.amazon.AmazonBackend.dto.ApiResponse;
-import com.phoenix.amazon.AmazonBackend.dto.PageableResponse;
-import com.phoenix.amazon.AmazonBackend.dto.PasswordUpdateDto;
-import com.phoenix.amazon.AmazonBackend.dto.UpdateUserDto;
-import com.phoenix.amazon.AmazonBackend.dto.UserDto;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.ApiResponse;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.PageableResponse;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.PasswordUpdateDto;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.UpdateUserDto;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.UserDto;
 import com.phoenix.amazon.AmazonBackend.exceptions.BadApiRequestExceptions;
 import com.phoenix.amazon.AmazonBackend.exceptions.UserExceptions;
 import com.phoenix.amazon.AmazonBackend.exceptions.UserNotFoundExceptions;
