+++ src/test/java/com/phoenix/amazon/AmazonBackend/controllers/UserControllerImplTest.java
@@ -2,18 +2,14 @@ package com.phoenix.amazon.AmazonBackend.controllers;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.phoenix.amazon.AmazonBackend.dto.ApiResponse;
-import com.phoenix.amazon.AmazonBackend.dto.ImageResponseMessages;
-import com.phoenix.amazon.AmazonBackend.dto.PageableResponse;
-import com.phoenix.amazon.AmazonBackend.dto.PasswordResponseMessages;
-import com.phoenix.amazon.AmazonBackend.dto.PasswordUpdateDto;
-import com.phoenix.amazon.AmazonBackend.dto.UserDto;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.PageableResponse;
+import com.phoenix.amazon.AmazonBackend.dto.responseDtos.PasswordResponseMessages;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.PasswordUpdateDto;
+import com.phoenix.amazon.AmazonBackend.dto.requestDtos.UserDto;
 
-import com.phoenix.amazon.AmazonBackend.services.IImageService;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.ClassOrderer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
